+++ YandexMobileAdsExample/app/src/androidTest/java/com/yandex/ads/sample/tests/RewardedLoadTest.kt
@@ -6,20 +6,17 @@ import com.yandex.ads.sample.HomeActivity
 import com.yandex.ads.sample.base.BaseUITest
 import com.yandex.ads.sample.base.defaultExt
 import com.yandex.ads.sample.pageobjects.RewardedScreen
-import com.yandex.ads.sample.pageobjects.clickCallToAction
-import com.yandex.ads.sample.pageobjects.clickShowAd
 import com.yandex.ads.sample.shared_steps.GoToSection
 import com.yandex.ads.sample.shared_steps.checkAnyBrowserOrStoreIsOpened
+import com.yandex.ads.sample.shared_steps.clickCallToAction
+import com.yandex.ads.sample.shared_steps.clickShowAd
 import com.yandex.ads.sample.shared_steps.choiceNetwork
 import com.yandex.ads.sample.shared_steps.goToSection
 import com.yandex.ads.sample.shared_steps.openSampleApp
 import com.yandex.mobile.ads.common.AdActivity
-import io.github.kakaocup.kakao.screen.Screen.Companion.onScreen
-import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
 
-@Ignore("Bad test stability")
 internal class RewardedLoadTest : BaseUITest() {
     @get:Rule
     val activityRule = activityScenarioRule<HomeActivity>()
@@ -45,25 +42,17 @@ internal class RewardedLoadTest : BaseUITest() {
             }
         }
 
-        step("Нажать на кнопку \"Show ad\"") {
-            onScreen<RewardedScreen> {
-                flakySafely {
-                    clickShowAd()
-                }
-            }
+        clickShowAd()
 
-            step("Успешно отобразилась реклама") {
-                flakySafely(60_000) {
-                    device.activities.isCurrent(AdActivity::class.java)
-                }
+        step("Успешно отобразилась реклама") {
+            flakySafely(60_000) {
+                device.activities.isCurrent(AdActivity::class.java)
             }
         }
 
-        step("Кликнуть по рекламе") {
-            onScreen<RewardedScreen> { clickCallToAction() }
+        clickCallToAction()
 
-            checkAnyBrowserOrStoreIsOpened("Выполнен корректный переход в браузер или на установку рекламируемого приложения")
-        }
+        checkAnyBrowserOrStoreIsOpened("Выполнен корректный переход в браузер или на установку рекламируемого приложения")
     }
 
     private companion object {