+++ YandexMobileAdsExample/app/src/main/java/com/yandex/ads/sample/adunits/RewardedAdActivity.kt
@@ -18,14 +18,14 @@ import com.yandex.ads.sample.network.Network
 import com.yandex.ads.sample.network.NetworkAdapter
 import com.yandex.ads.sample.utils.applySystemBarsPadding
 import com.yandex.mobile.ads.common.AdError
-import com.yandex.mobile.ads.common.AdRequestConfiguration
+import com.yandex.mobile.ads.common.AdRequest
 import com.yandex.mobile.ads.common.AdRequestError
 import com.yandex.mobile.ads.common.ImpressionData
 import com.yandex.mobile.ads.rewarded.Reward
 import com.yandex.mobile.ads.rewarded.RewardedAd
 import com.yandex.mobile.ads.rewarded.RewardedAdEventListener
-import com.yandex.mobile.ads.rewarded.RewardedAdLoader
 import com.yandex.mobile.ads.rewarded.RewardedAdLoadListener
+import com.yandex.mobile.ads.rewarded.RewardedAdLoader
 
 class RewardedAdActivity : AppCompatActivity(R.layout.activity_rewarded_ad),
     RewardedAdLoadListener {
@@ -51,19 +51,15 @@ class RewardedAdActivity : AppCompatActivity(R.layout.activity_rewarded_ad),
         // Initialize SDK as early as possible, for example in Application.onCreate or at least Activity.onCreate
         // It's recommended to use the same instance of InterstitialAdLoader for every load for
         // achieve better performance
-        rewardedAdLoader = RewardedAdLoader(this).apply {
-            setAdLoadListener(this@RewardedAdActivity)
-        }
+        rewardedAdLoader = RewardedAdLoader(this)
         loadRewardedAd()
     }
 
     private fun loadRewardedAd() {
         disabeShowAdButton()
-        rewardedAdLoader?.loadAd(createAdRequestConfiguration())
+        val adRequest = AdRequest.Builder(selectedNetwork.adUnitId).build()
+        rewardedAdLoader?.loadAd(adRequest, this)
     }
-
-    private fun createAdRequestConfiguration(): AdRequestConfiguration =
-        AdRequestConfiguration.Builder(selectedNetwork.adUnitId).build()
 
     private fun showRewardedAd() {
         disabeShowAdButton()
@@ -94,8 +90,7 @@ class RewardedAdActivity : AppCompatActivity(R.layout.activity_rewarded_ad),
     }
 
     override fun onDestroy() {
-        // Set listener to null to avoid memory leaks
-        rewardedAdLoader?.setAdLoadListener(null)
+        rewardedAdLoader?.cancelLoading()
         rewardedAdLoader = null
         destroyRewardedAd()
         super.onDestroy()
@@ -160,20 +155,52 @@ class RewardedAdActivity : AppCompatActivity(R.layout.activity_rewarded_ad),
 
         private val networks = arrayListOf(
             Network(R.drawable.ic_yandex_icon_24, R.string.yandex_title, "demo-rewarded-yandex"),
-            Network(R.drawable.ic_admanager_icon, R.string.admanager_title, "demo-rewarded-admanager"),
+            Network(
+                R.drawable.ic_admanager_icon,
+                R.string.admanager_title,
+                "demo-rewarded-admanager"
+            ),
             Network(R.drawable.ic_admob_icon_24, R.string.admob_title, "demo-rewarded-admob"),
-            Network(R.drawable.ic_applovin_icon_24, R.string.applovin_title, "demo-rewarded-applovin"),
+            Network(
+                R.drawable.ic_applovin_icon_24,
+                R.string.applovin_title,
+                "demo-rewarded-applovin"
+            ),
             Network(R.drawable.ic_appnext_icon, R.string.appnext_title, "demo-rewarded-appnext"),
             Network(R.drawable.ic_bigoads_icon, R.string.bigoads_title, "demo-rewarded-bigoads"),
-            Network(R.drawable.ic_chartboost_icon, R.string.chartboost_title, "demo-rewarded-chartboost"),
+            Network(
+                R.drawable.ic_chartboost_icon,
+                R.string.chartboost_title,
+                "demo-rewarded-chartboost"
+            ),
             Network(R.drawable.ic_inmobi_icon, R.string.inmobi_title, "demo-rewarded-inmobi"),
-            Network(R.drawable.ic_mintegral_logo, R.string.mintegral_title, "demo-rewarded-mintegral"),
-            Network(R.drawable.ic_mytarget_icon_24, R.string.my_target_title, "demo-rewarded-mytarget"),
-            Network(R.drawable.ic_ironsource_icon_24, R.string.ironsource_title, "demo-rewarded-ironsource"),
+            Network(
+                R.drawable.ic_mintegral_logo,
+                R.string.mintegral_title,
+                "demo-rewarded-mintegral"
+            ),
+            Network(
+                R.drawable.ic_mytarget_icon_24,
+                R.string.my_target_title,
+                "demo-rewarded-mytarget"
+            ),
+            Network(
+                R.drawable.ic_ironsource_icon_24,
+                R.string.ironsource_title,
+                "demo-rewarded-ironsource"
+            ),
             Network(R.drawable.ic_pangle_icon, R.string.pangle_title, "demo-rewarded-pangle"),
-            Network(R.drawable.ic_startapp_icon_24, R.string.startapp_title, "demo-rewarded-startapp"),
+            Network(
+                R.drawable.ic_startapp_icon_24,
+                R.string.startapp_title,
+                "demo-rewarded-startapp"
+            ),
             Network(R.drawable.ic_tapjoy_icon, R.string.tapjoy_title, "demo-rewarded-tapjoy"),
-            Network(R.drawable.ic_unityads_icon_24, R.string.unity_ads_title, "demo-rewarded-unityads"),
+            Network(
+                R.drawable.ic_unityads_icon_24,
+                R.string.unity_ads_title,
+                "demo-rewarded-unityads"
+            ),
             Network(R.drawable.ic_vungle_icon_24, R.string.vungle_title, "demo-rewarded-vungle"),
         )
     }