+++ changelogs/mediation/applovin/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to AppLovin Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 13.5.1.0-beta.1
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.0.0-beta.1
+* Applovin SDK version 13.5.1
+
 ## Version 13.1.0.12
 
 #### Updated