+++ changelogs/mediation/mytarget/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to MyTarget Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 	5.45.3.1-beta.1
+
+#### Breaking Changes
+* Raised the Android minSdk version to 23
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.0.0-beta.1
+* Mytarget SDK version 5.45.3
+
 ## Version 5.27.4.1
 
 #### Updated