+++ changelogs/mediation/petalads/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Change Log
+All notable changes to PetalAds Adapter for Yandex Mobile Ads Mediation will be documented in this file.
+
+## Version 3.4.79.301.7-beta.1
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.0.0-beta.1
+
 ## Version 3.4.79.301.6
 
 #### Updated